Trocado para Integer

+++ tickerscraper/src/main/java/br/dev/rodrigopinheiro/tickerscraper/infrastructure/scraper/fii/dto/FiiDadosFinanceirosDTO.java
@@ -4,7 +4,7 @@ package br.dev.rodrigopinheiro.tickerscraper.infrastructure.scraper.fii.dto;
 import java.util.List;
 
 public record FiiDadosFinanceirosDTO(
-        Long internalId,
+        Integer internalId,
         FiiInfoHeaderDTO infoHeader,
         FiiIndicadorHistoricoDTO infoHistorico,
         FiiInfoSobreDTO infoSobre,
